make briefing tabs wider
